refactor (ride-service): reorganize DTOs into request and response packages
