+++ ride-service/src/main/java/com/software/modsen/rideservice/controller/RideApi.java
@@ -1,8 +1,8 @@
 package com.software.modsen.rideservice.controller;
 
-import com.software.modsen.rideservice.dto.RideListDto;
-import com.software.modsen.rideservice.dto.RideRequest;
-import com.software.modsen.rideservice.dto.RideResponse;
+import com.software.modsen.rideservice.dto.response.RideListDto;
+import com.software.modsen.rideservice.dto.request.RideRequest;
+import com.software.modsen.rideservice.dto.response.RideResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
