+++ ride-service/src/main/java/com/software/modsen/rideservice/mapper/RideMapper.java
@@ -1,7 +1,7 @@
 package com.software.modsen.rideservice.mapper;
 
-import com.software.modsen.rideservice.dto.RideRequest;
-import com.software.modsen.rideservice.dto.RideResponse;
+import com.software.modsen.rideservice.dto.request.RideRequest;
+import com.software.modsen.rideservice.dto.response.RideResponse;
 import com.software.modsen.rideservice.model.Ride;
 import org.mapstruct.Mapper;
 import org.mapstruct.MappingTarget;
