+++ ride-service/src/main/java/com/software/modsen/rideservice/service/RideService.java
@@ -1,8 +1,8 @@
 package com.software.modsen.rideservice.service;
 
-import com.software.modsen.rideservice.dto.RideListDto;
-import com.software.modsen.rideservice.dto.RideRequest;
-import com.software.modsen.rideservice.dto.RideResponse;
+import com.software.modsen.rideservice.dto.response.RideListDto;
+import com.software.modsen.rideservice.dto.request.RideRequest;
+import com.software.modsen.rideservice.dto.response.RideResponse;
 
 public interface RideService {
     RideListDto getRides();
