+++ ride-service/src/main/java/com/software/modsen/rideservice/service/impl/DefaultRideService.java
@@ -1,8 +1,8 @@
 package com.software.modsen.rideservice.service.impl;
 
-import com.software.modsen.rideservice.dto.RideListDto;
-import com.software.modsen.rideservice.dto.RideRequest;
-import com.software.modsen.rideservice.dto.RideResponse;
+import com.software.modsen.rideservice.dto.response.RideListDto;
+import com.software.modsen.rideservice.dto.request.RideRequest;
+import com.software.modsen.rideservice.dto.response.RideResponse;
 import com.software.modsen.rideservice.exception.InvalidResourceException;
 import com.software.modsen.rideservice.exception.ResourceNotFoundException;
 import com.software.modsen.rideservice.mapper.RideMapper;
@@ -13,7 +13,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
