+++ ride-service/src/test/java/com/software/modsen/rideservice/controller/RideControllerTest.java
@@ -2,8 +2,8 @@ package com.software.modsen.rideservice.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.software.modsen.rideservice.controller.impl.RideController;
-import com.software.modsen.rideservice.dto.RideListDto;
-import com.software.modsen.rideservice.dto.RideRequest;
+import com.software.modsen.rideservice.dto.response.RideListDto;
+import com.software.modsen.rideservice.dto.request.RideRequest;
 import com.software.modsen.rideservice.service.RideService;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
