+++ ride-service/src/test/java/com/software/modsen/rideservice/util/RideTestUtil.java
@@ -1,7 +1,7 @@
 package com.software.modsen.rideservice.util;
 
-import com.software.modsen.rideservice.dto.RideRequest;
-import com.software.modsen.rideservice.dto.RideResponse;
+import com.software.modsen.rideservice.dto.request.RideRequest;
+import com.software.modsen.rideservice.dto.response.RideResponse;
 import com.software.modsen.rideservice.model.Ride;
 import com.software.modsen.rideservice.model.enums.Status;
 import lombok.experimental.UtilityClass;
